Validation: created validation for categories and products.

+++ app/Http/Controllers/Admin/Product/ProductController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\Category;
 use Illuminate\Http\Request;
+use App\Http\Requests\ProductsRequest;
 use Illuminate\Support\Facades\Storage;
 
 class ProductController extends Controller
@@ -40,11 +41,14 @@ class ProductController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(ProductsRequest $request)
     {
-        $path = $request->file('image')->store('public/products');
         $params = $request->all();
-        $params['image'] = $path;
+        if($request->has('image')){
+            $path = $request->file('image')->store('public/products');
+            $params['image'] = $path;
+        }
+
         Product::create($params);
     
         return redirect()->route('products.index');
@@ -81,12 +85,16 @@ class ProductController extends Controller
      * @param  \App\Models\Product  $product
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Product $product)
+    public function update(ProductsRequest $request, Product $product)
     {
-        Storage::delete($product->image);
-        $path = $request->file('image')->store('public/products');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if($request->has('image')){
+            Storage::delete($product->image);
+            $path = $request->file('image')->store('public/products');
+            $params['image'] = $path;
+        }
+
         $product->update($params);
         
         return redirect()->route('products.index');
